Generation of Sphere!
The sphere mesh generation is completed! Though we need to address the color situation (the URP materials, yep, it would've been better to dive into the 3D normal pipeline) and the small bumps in the sphere due to the difference in the distance of each vertice in the initial cube to its center.